+++ lib/currency_widget.dart
@@ -9,3 +9,4 @@ export 'package:currency_widget/src/view/currency_chooser.dart';
 export 'package:currency_widget/src/view/custom_currency_chooser.dart';
 export 'package:currency_widget/src/view/custom_currency_picker.dart';
 export 'package:currency_widget/src/view/currency_multi_selector.dart';
+export 'package:currency_widget/src/view/custom_currency_dropdown.dart';